Update Home, CreatePostButton, and PostForm components
Home.js is updated to include "Button" in the imports and the axios call is switched from "get" to "post". A new structure and visuals for posts are also introduced. The CreatePostButton.js has changes in headers while PostForm.js has a whole new structure now for file upload and form submission.