Simplify conditional rendering in TopBar component
The code for the TopBar component has been updated to simplify the conditional rendering of the ButtonContainer. Previously, the code was using a ternary operator for this, but since the else clause was eliminated, the conditional (&&) operator is a more appropriate choice. This offers a cleaner and more efficient way of rendering the ButtonContainer, without the need for an unnecessary else case.